faithful orthodox address arithmetic

+++ srcs/mhsize.c
@@ -7,5 +7,6 @@
 #undef LIBM_INTERNAL
 
 inline size_t	mhsize(void) {
-	return ((!__mstart || !__mend) ? (size_t)0 : (size_t)(__mend - __mstart));
+	return ((!__mstart || !__mend)
+		? (size_t)0 : (size_t)((uintptr_t)__mend - (uintptr_t)__mstart));
 }
